# Fixed several performance issues and stuttering