--- a/src/main/kotlin/org/acejump/control/Trigger.kt
+++ b/src/main/kotlin/org/acejump/control/Trigger.kt
@@ -2,47 +2,40 @@ package org.acejump.control
 
 import com.intellij.openapi.diagnostic.Logger
 import org.acejump.search.runLater
-import java.lang.System.currentTimeMillis
-import kotlin.math.abs
+import java.util.concurrent.Executors
+import java.util.concurrent.Future
+import java.util.concurrent.TimeUnit
 
 /**
  * Timer for triggering events with a designated delay.
  */
 
-object Trigger: () -> Unit {
-  private val logger = Logger.getInstance(Trigger::class.java)
-  private var delay = 0L
-  private var timer = currentTimeMillis()
-  private var isRunning = false
-  private var invokable: () -> Unit = {}
-
-  override fun invoke() {
-    timer = currentTimeMillis()
-    if (isRunning) return
-    synchronized(this) {
-      isRunning = true
-
-      while (currentTimeMillis() - timer <= delay)
-        Thread.sleep(abs(delay - (currentTimeMillis() - timer)))
-
-      try {
-        invokable()
-      } catch (e: Exception) {
-        logger.error("Exception occurred while triggering event!", e)
-      }
-
-      isRunning = false
-    }
+class Trigger {
+  private companion object {
+    private val executor = Executors.newSingleThreadScheduledExecutor()
   }
+
+  private val logger = Logger.getInstance(Trigger::class.java)
+  private var task: Future<*>? = null
 
   /**
    * Can be called multiple times inside [delay], but doing so will reset the
    * timer, delaying the [event] from occurring by [withDelay] milliseconds.
    */
 
-  operator fun invoke(withDelay: Long = 750, event: () -> Unit = {}) {
-    delay = withDelay
-    invokable = event
-    runLater(this)
+  @Synchronized
+  operator fun invoke(withDelay: Long, event: () -> Unit = {}) {
+    task?.cancel(true)
+    task = executor.schedule({
+      runLater {
+        try {
+          event()
+        } catch (e: Exception) {
+          logger.error("Exception occurred while triggering event!", e)
+        } finally {
+          task = null
+        }
+      }
+    }, withDelay, TimeUnit.MILLISECONDS)
   }
 }