Was trying the plugin out and immediately ran into severe performance issues, so I figured I'd give some optimizations a shot.

Here are my results with the `test tag latency` test:
```
Before: 25.1s (avg 185ms)
After:   9.2s (avg  28ms)
```

I verified that unit tests succeed (apart from `test pinyin selection` which had a compile error in `master`, not sure if you know), and that the order of tags has not changed and basic uses of the plugin still work, but since I've only used it for a few minutes so far it's possible something subtle may have changed that I didn't catch.

### Rationale

In a few places I avoided recomputing things that don't change, those cases should be obvious.

In `Solver.tryToAssignTag`, the first major change was storing used indices in an `IntSet` which already brought it down from about 24s to about 17s. HashMap value collections are O(n), plus `IntSet` has the benefit of not boxing primitives. I also changed `newTags` into an `Object2IntOpenHashMap` to avoid boxing values. These collections come from fastutil included in IJ.

In `Solver.map`, I completely rewrote the collections and removed parallel streams. Regarding parallelization, I have a 16 thread CPU and the overhead of starting threads and synchronizing on the `TreeMultimap` was so large it took all the threads just to match the performance of single-threaded streams. Reducing the thread count made performance worse, so this parallel stream likely runs worse on any CPU with fewer than 16 threads than a non-parallel stream.

Still, the major bottleneck was `TreeMultimap`. I found that the values only needed to be in a sorted state by the time they get passed to `tryToAssignTag`, so I completely replaced the multimap with a `HashMap<String, IntList>` where `IntList` not only avoids boxing, but also supports primitive integer comparators for sorting, and has an efficient conversion to array.

Switching out the map and removing parallel streams brought it down from 17s to 11s. Replacing the boxed comparator with primitive comparator made the code a bit more verbose, but it also saved roughly 2 additional seconds. I noticed the collections end up rather large and this is probably close to the limit of what can be done without completely rewriting how tags are assigned.

In `AceUtil.wordBoundsPlus`, I turned the sequence into a simple loop, it didn't save much time but I didn't see any reason to use a sequence and allocate collections there.

Finally, the old implementation of `Trigger` appears to have been freezing the UI thread and didn't work as described in the documentation comment (it wasn't resetting the timer on repeated calls despite the comment saying so). My reimplementation probably isn't the best either, but with it and with the other optimizations, I was able to drop the `visibleAreaChanged` delay to about half with no stuttering.