@chylex 2020-11-18
I found a few more opportunities, at this point the included test is running too fast to have any meaningful data so I bumped up the repeats to 800 and got
```
Before: 34s (avg 22.8ms)
After:  28s (avg 15.6ms)
```
(might be a good idea to do a warmup in the tests, bumping up the repeats reduced averages overall)

I think it's pretty tight at this point, the next meaningful step could be only searching inside visible area and then lazily adding new markers as the user navigates past the top/bottom marker, but that'd be more work than what I want to put into it.