@breandan 2020-11-20
Hi @chylex, thanks for the PR, this is really solid work! I was able to reproduce the latency reduction on some texts, although the variance is still pretty high depending on the query and editor contents. I also added a lorem ipsum test in c8c6f9c and noticed an improvement, but did not notice a significant improvement on random text. Most of the optimizations make sense, although a more comprehensive set of performance tests are needed. Ideally, these would include a representative sample of random source code files.

I am running `./gradlew test -i` on `IU-2020.3` running on a `Intel(R) Xeon(R) CPU E3-1505M v5 @ 2.80GHz`.

These were the results I collected before merging this PR as of 6c19767:

```
AceTest > test lorem ipsum latency STANDARD_OUT
    Average time to tag results: 742ms

AceTest > test random text latency STANDARD_OUT
    Average time to tag results: 1403ms
```

These were the results I observed after merging this PR as of a38c670:

```
AceTest > test lorem ipsum latency STANDARD_OUT
    Average time to tag results: 72ms

AceTest > test random text latency STANDARD_OUT
    Average time to tag results: 1322ms
```

> the next meaningful step could be only searching inside visible area and then lazily adding new markers as the user navigates past the top/bottom marker

Assigning text by only considering the visible area has been suggested, although this would not be compatible with AceJump's current search functionality due to the [tag assignment problem](https://github.com/acejump/AceJump/blob/7e4168da76b3810c9fba3c8470815f1a643fe512/src/main/kotlin/org/acejump/label/Solver.kt#L13). In order to assign tags, we must ensure no tags could ever result in a collision, which is basically an ambiguous string elsewhere in the editor text. We do not currently test for tag collisions, and I did not check very carefully, but it looks like your PR does not break this assumption.

> I'll continue and probably make another PR tomorrow

I have merged this was planning to release in `3.6.4`, but if you prefer, I will wait for your next PR.

> unit tests succeed (apart from `test pinyin selection` which had a compile error in `master`, not sure if you know),

I could not reproduce this error. Please open an issue if you happen to encounter it again.