@chylex 2020-11-20
> > unit tests succeed (apart from test pinyin selection which had a compile error in master, not sure if you know),

> I could not reproduce this error. Please open an issue if you happen to encounter it again.

It's strange, I tried loading the project on a linux system with IJ 2020.2.3 and it compiles there, but my main windows system with 2020.3 EAP (203.5251.39) complains:
![obrazek](https://user-images.githubusercontent.com/3685160/99775361-773e2f80-2b0f-11eb-9091-59eba78d04fd.png)

> I have merged this was planning to release in 3.6.4, but if you prefer, I will wait for your next PR.

I'll make another PR later today, so you could wait since it'll probably take until next week for JetBrains to review the update. I also added a few new features and improvements, but I'll open an issue with more details and ask you which features you'd like me to PR separately from optimizations.